Remove the json_schema dependency
We do not want to include the json_schema as a dependency of the gem, since the
idea is to the json_schema to be used only by the test helper.

This patch load the json_schema only inside the `assert_response_schema` test
helper and raises a good error message when the the json_schema gem is not
installed.